Extracted to ``SearchBar.tsx``
